+++ lib/ide/Cargo.toml
@@ -9,11 +9,12 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 basegl                   = { version = "0.1.0"  , path = "../core"          }
+basegl-core-msdf-sys     = { version = "0.1.0"  , path = "../core/msdf-sys" }
+basegl-system-web        = { version = "0.1.0"  , path = "../system/web"    }
 enso-prelude             = { version = "0.1.0"  , path = "../prelude"       }
 file-manager-client      = { version = "0.1.0"  , path = "file-manager"     }
 json-rpc                 = { version = "0.1.0"  , path = "json-rpc"         }
 utils                    = { version = "0.1.0"  , path = "utils"            }
-basegl-core-msdf-sys     = { version = "0.1.0"  , path = "../core/msdf-sys" }
 shapely                  = { version = "0.1.0"  , path = "../shapely/impl"  }
 
 console_error_panic_hook = { version = "0.1.6"                           }
