+++ lib/ide/src/transport/web.rs
@@ -2,13 +2,11 @@
 
 use crate::prelude::*;
 
+use basegl_system_web::closure::ClosureStorage;
 use failure::Error;
 use futures::channel::mpsc;
-use js_sys::Function;
 use json_rpc::Transport;
 use json_rpc::TransportEvent;
-use wasm_bindgen::convert::FromWasmAbi;
-use wasm_bindgen::JsCast;
 use web_sys::CloseEvent;
 use web_sys::Event;
 use web_sys::MessageEvent;
@@ -97,60 +95,6 @@ impl State {
             web_sys::WebSocket::CLOSED     => State::Closed,
             num                            => State::Unknown(num), // impossible
         }
-    }
-}
-
-
-
-// ======================
-// === ClosureStorage ===
-// ======================
-
-/// Constraint for JS closure argument types
-pub trait ClosureArg = FromWasmAbi + 'static;
-
-/// Function that can be wrapped into a `Closure`.
-pub trait ClosureFn<Arg> = FnMut(Arg) + 'static where Arg: ClosureArg;
-
-/// Stores an optional closure.
-#[derive(Debug,Derivative)]
-#[derivative(Default(bound=""))]
-pub struct ClosureStorage<Arg> {
-    /// The stored closure.
-    pub closure : Option<Closure<dyn FnMut(Arg)>>,
-}
-
-impl <Arg> ClosureStorage<Arg> {
-    /// An empty closure storage.
-    pub fn new() -> ClosureStorage<Arg> {
-        default()
-    }
-
-    /// Stores the given closure.
-    pub fn store(&mut self, closure:Closure<dyn FnMut(Arg)>) {
-        self.closure = Some(closure);
-    }
-
-    /// Obtain JS reference to the closure (that can be passed e.g. as a callback
-    /// to an event handler).
-    pub fn js_ref(&self) -> Option<&Function> {
-        self.closure.as_ref().map(|closure| closure.as_ref().unchecked_ref())
-    }
-
-    /// Wraps given function into a Closure.
-    pub fn wrap(&mut self, f:impl ClosureFn<Arg>) {
-        let boxed = Box::new(f);
-        // Note: [mwu] Not sure exactly why, but compiler sometimes require this
-        // explicit type below and sometimes does not.
-        let wrapped:Closure<dyn FnMut(Arg)> = Closure::wrap(boxed);
-        self.store(wrapped);
-    }
-
-    /// Clears the current closure.
-    /// Note: if reference to it is still used by JS, it will throw an exception
-    /// on calling attempt. Be careful of dangling references.
-    pub fn clear(&mut self) {
-        self.closure = None;
     }
 }
